Support --key in MVT/postserve/test for MD5 hash
The --key generates MVT queries with an extra column
that contains the md5 hash of the generated tile.

Note that it will not be identical to the one generated
by mbtiles in order to detect duplicates, because
mbtiles generates it after gziping, whereas here
PostgreSQL has no gzip() function (yet?), so md5()
is done on the raw tile.