Consolidate VERSION to a single place
There were two places for the same version info,
in the VERSION file in the root, and the __init__.py,
and they got out of sync.

VERSION is easier to maintain, but it becomes tricky
to use it from Python, so now it instead uses
the __init__.py version for everything including the makefiles.

A few more minor fixes:
 - reuse build targets $@ in the makefile