SQL importing and env var improvements
### environment variable cleanup

Long overdue cleanup of the environment variables to clarify their meaning.  Ideally we should do this as part of import-osm -> tools migration because it requires people to update their docker image usage anyway, so would be good to modify it at the same time.

* rename `CONFIG_JSON` ⇒ `IMPOSM_CONFIG_FILE`
* rename `MAPPING_YAML` ⇒ `IMPOSM_MAPPING_FILE`
* rename `IMPORT_DIR` ⇒ `PBF_DATA_DIR`
* rename `DIFF_DIR` ⇒ `IMPOSM_DIFF_DIR`
* rename `TILES_DIR` ⇒ `EXPIRETILES_DIR`

* rename `OMT_UTIL_DIR` and `VT_UTIL_DIR` ⇒ `SQL_TOOLS_DIR` (merged)

### related improvements
* Combine language and utility SQL files into a single dir - `/sql`, which means both `VT_UTIL_DIR` and `OMT_UTIL_DIR` have been removed, and a single `SQL_TOOLS_DIR` location was added.
* `import-osm` can now accept PBF file as the first parameter, rather than auto-search for the first file in the `PBF_DATA_DIR`

A minor side change -- validate that the variables are set and abort early