v5.2 automate imposm10 docker image publishing (#281)
Github action will publish:
* on `master` branch push -- `latest` and `latest-imposm10`
* on `vN.N` branch push -- `test-N.N` (no Imposm)
* on `vN.N.N` **TAG** push -- `N.N.N, N.N, N.N.N-imposm10, N.N-imposm10`

The versions will be pushblish for ALL images, but in reality the only different image will be the tools. All other images (such as postgis, preload, etc) will have the same content.  This is done to simplify OMT usage - just set `TOOLS_VERSION=N.N or N.N-imposm10`, and it will work.

To support the above, `DOCKER_BUILD_EXTRAS` param was added to the Makefile.