New layer-stats tool to explore data (#293)
The tool is modeled on the OpenMapTiles `qa/` tools. The tool shows some statistics for a given layer's field for each zoom. It currently supports 3 methods:

* `frequency` -- Shows how often each unique value occurs in a layer's column. If more than one column is given, shows unique combinations.
* `toplength` -- Shows the longest N values for a given layer's column.
* `variance` -- Shows a few statistical metrics for a column's numeric value.

This is a simple test script I used outside of the docker, and it could be adapted to run inside the docker as well. https://gist.github.com/nyurik/457102422adfb380ec9e8552d7656112