Use print_err in places that mix errors and output on stdout. (#307)
We use these scripts in CI and since most of the scripts push results to `stdout`, we redirect the result to a file. This works great, until there's an error.

When there's an error, a few of the scripts push the errors to `stdout` instead of `stderr`. This means it appears the commands succeeded when they actually failed.

This commit adjusts some `print` statements to use `print_err` to ensure these results go to `stderr`.